streams: moved featured communities into communities.
