+++ invenio_rdm_migrator/streams/communities/table_generator.py
@@ -6,15 +6,39 @@
 # it under the terms of the MIT License; see LICENSE file for more details.
 
 """Invenio RDM migration user table load module."""
+import random
 
 from ...load.postgresql import TableGenerator, generate_uuid
-from .models import Community, CommunityMember
+from .models import Community, CommunityMember, FeaturedCommunity
 
 
 def _generate_members_uuids(data):
     for member in data["community_members"]:
         member["id"] = generate_uuid(data)
     return data["community_members"]
+
+
+def _generate_featured_community_id(data):
+    if not data.get("featured_community"):
+        return None
+    return _pid_pk(data)
+
+
+# keep track of generated PKs, since there's a chance they collide
+GENERATED_PID_PKS = set()
+
+
+def _pid_pk(data):
+    """Generates a random integer between 1M and 2B, avoiding duplication with previously generated numbers."""
+    lower_limit = 1_000_000
+    upper_limit = 2_147_483_647 - 1
+
+    while True:
+        # we start at 1M to avoid collisions with existing low-numbered PKs
+        val = random.randint(lower_limit, upper_limit)
+        if val not in GENERATED_PID_PKS:
+            GENERATED_PID_PKS.add(val)
+            return val
 
 
 class CommunityTableGenerator(TableGenerator):
@@ -24,10 +48,11 @@ class CommunityTableGenerator(TableGenerator):
         """Constructor."""
         self.communities_cache = communities_cache
         super().__init__(
-            tables=[Community, CommunityMember],
+            tables=[Community, CommunityMember, FeaturedCommunity],
             pks=[
                 ("community.id", generate_uuid),
                 ("community_members", _generate_members_uuids),
+                ("featured_community.id", _generate_featured_community_id),
             ],
         )
 
@@ -44,6 +69,11 @@ class CommunityTableGenerator(TableGenerator):
         for member in community_members:
             member["community_id"] = community_id
             yield CommunityMember(**member)
+
+        featured_community = data["featured_community"]
+        if featured_community["id"]:
+            featured_community["community_id"] = community_id
+            yield FeaturedCommunity(**featured_community)
 
     def cleanup(self, **kwargs):
         """Cleanup."""
